Mac ARM64 integration
* Update Mac platform to support multi-arch detection
* New Mac arm64 specific cmake files
* Remove EditorCommon.Headers causing issues with Qt6 moving and Mac

* Fix QAbstractNativeEventFilter::nativeEventFilter
 Signature
* Update deprecated constructor call to QMouseEvent
* Update deprecated QHoverEvent->pos() to ->position)
* Fix issue where file that needs mocing 'QtEditorApplication_mac.h' was not in any cmake file list, thus not being auto-moc'd
* Removed unneeded include for that moc
* Added static_cast<int> in numerous places to fix error "Implicit conversion loses integer precision: 'qsizetype' (aka 'long long') to 'int'"

+++ Code/Editor/Controls/ConsoleSCB.cpp
@@ -78,14 +78,14 @@ protected:
 
         while (true)
         {
-            pos = text.indexOf(m_searchTerm, pos+1, Qt::CaseInsensitive);
+            pos = static_cast<int>(text.indexOf(m_searchTerm, pos+1, Qt::CaseInsensitive));
 
             if (pos == -1)
             {
                 break;
             }
 
-            setFormat(pos, m_searchTerm.length(), myClassFormat);
+            setFormat(pos, static_cast<int>(m_searchTerm.length()), myClassFormat);
         }
     }
 
@@ -112,7 +112,7 @@ static QString RemoveColorCode(const QString& text, int& iColorCode)
     QString cleanString;
     cleanString.reserve(text.size());
 
-    const int textSize = text.size();
+    const int textSize = static_cast<int>(text.size());
     for (int i = 0; i < textSize; ++i)
     {
         QChar c = text.at(i);
@@ -296,7 +296,7 @@ void ConsoleLineEdit::DisplayHistory(bool bForward)
 
 void ConsoleLineEdit::ResetHistoryIndex()
 {
-    m_historyIndex = m_history.size();
+    m_historyIndex = static_cast<int>(m_history.size());
     m_bReusedHistory = false;
 }
 
@@ -705,11 +705,11 @@ void ConsoleVariableItemDelegate::setEditorData(QWidget* editor, const QModelInd
         Q_ASSERT(var->GetType() == IVariable::FLOAT);
 
         QString valStr = QString::number(value.toFloat());
-        int decimalIndex = valStr.indexOf('.');
+        int decimalIndex = static_cast<int>(valStr.indexOf('.'));
         if (decimalIndex != -1)
         {
             valStr.remove(0, decimalIndex + 1);
-            doubleEditor->setDecimals(valStr.size());
+            doubleEditor->setDecimals(static_cast<int>(valStr.size()));
         }
 
         // Set the initial value to our editor
