+++ Gems/EMotionFX/Code/Source/Editor/Plugins/ColliderWidgets/ClothOutlinerNotificationHandler.cpp
@@ -69,7 +69,7 @@ namespace EMotionFX
             return;
         }
 
-        const int numSelectedJoints = selectedRowIndices.count();
+        const int numSelectedJoints = static_cast<int>(selectedRowIndices.count());
         int numJointsWithColliders = 0;
         for (const QModelIndex& modelIndex : selectedRowIndices)
         {
